Add SPDK BlobFS integration.
This patches adds support for using the SPDK BlobFS as the
storage engine for a RocksDB database.

The Makefile changes assume that SPDK and DPDK have been
built according to the SPDK README (https://github/spdk/spdk)
and share the same parent directory as the RocksDB repository.

Flags to configure the SPDK storage engine:

spdk -
  Defines the name of the SPDK configuration file.  If
  omitted, will use PosixEnv instead of SpdkEnv.  This
  configuration file defines the block devices that may
  contain the BlobFS for the database.

spdk_bdev -
  Defines the name of the SPDK block device which
  contains the BlobFS for this database.

spdk_cache_size -
  Defines the amount of userspace cache memory used by
  BlobFS.  Specified in terms of megabytes (MB).  Default
  is 4096 (4GB).

Change-Id: I72edabc27191492b1d1b6a3e314390c803251571
Signed-off-by: Jim Harris <james.r.harris@intel.com>
Signed-off-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-on: https://review.gerrithub.io/373639